fix displaying the summary in Firefox
In FF, jQuery already parses the JSON for us when using get:

```
> $.get("/events/"+get_current_event_id()+"/summary", function(data) { console.log(data); })
Object { summary: "It was full of stars." }
```

In Chrome, it does not:

```
> $.get("/events/"+get_current_event_id()+"/summary", function(data) { console.log(data); })
{"summary":"It was full of stars."}
```

Therefore, parsing the "JSON" (actually, the string representation of
the already parsed object) fails and the summary does not display.

Use `getJSON` to let jQuery deal with this for us.